+++ snapshot-tests-core/src/main/java/de/skuzzle/test/snapshots/directories/DirectoryResolver.java
@@ -8,7 +8,7 @@ public class DirectoryResolver {
     // This will not necessarily resolve to the correct directory. Especially multi
     // module setups and cases where the build is not invoked directly from its folder,
     // this might give some problems
-    private final static Path BASE = Path.of("src/test/resources");
+    private final static Path BASE = Path.of("src", "test", "resources");
 
     public static Path resolve(String directory) throws IOException {
         return BASE.resolve(directory);