+++ snapshot-tests-core/src/main/java/de/skuzzle/test/snapshots/impl/SnapshotConfiguration.java
@@ -7,12 +7,12 @@ import java.util.Objects;
 
 import org.junit.jupiter.api.extension.ExtensionContext;
 
-import de.skuzzle.test.snapshots.SnapshotAssertions;
+import de.skuzzle.test.snapshots.EnableSnapshotTests;
 import de.skuzzle.test.snapshots.directories.DirectoryResolver;
 
 /**
  * Relevant configuration for executing snapshot tests in a test class that is annotated
- * with {@link SnapshotAssertions}.
+ * with {@link EnableSnapshotTests}.
  *
  * @author Simon Taddiken
  */
@@ -29,16 +29,16 @@ final class SnapshotConfiguration {
     }
 
     public Path determineSnapshotDirectory() throws IOException {
-        final String testDirName = snapshotDirecotry();
+        final String testDirName = snapshotDirecotryName();
 
         final Path testDirectory = DirectoryResolver.resolve(testDirName);
         Files.createDirectories(testDirectory);
         return testDirectory;
     }
 
-    private String snapshotDirecotry() throws IOException {
-        final SnapshotAssertions snapshotAssertions = extensionContext.getRequiredTestClass()
-                .getAnnotation(SnapshotAssertions.class);
+    private String snapshotDirecotryName() throws IOException {
+        final EnableSnapshotTests snapshotAssertions = extensionContext.getRequiredTestClass()
+                .getAnnotation(EnableSnapshotTests.class);
 
         final String testDirName = snapshotAssertions.snapshotDirectory().isEmpty()
                 ? extensionContext.getRequiredTestClass().getName().replace('.', '/') + "_snapshots"
@@ -52,13 +52,13 @@ final class SnapshotConfiguration {
 
     public boolean isForceUpdateSnapshots() {
         return extensionContext.getRequiredTestClass()
-                .getAnnotation(SnapshotAssertions.class)
+                .getAnnotation(EnableSnapshotTests.class)
                 .forceUpdateSnapshots();
     }
 
     public boolean isSoftAssertions() {
         return extensionContext.getRequiredTestClass()
-                .getAnnotation(SnapshotAssertions.class)
+                .getAnnotation(EnableSnapshotTests.class)
                 .softAssertions();
     }
 }