+++ snapshot-tests-jackson/src/test/java/de/skuzzle/test/snapshots/SnapshotsTest.java
@@ -8,8 +8,9 @@ import java.time.LocalDate;
 import org.junit.jupiter.api.Test;
 
 import de.skuzzle.test.snapshots.SnapshotDsl.Snapshot;
+import de.skuzzle.test.snapshots.SnapshotTestResult.SnapshotStatus;
 
-@SnapshotAssertions(forceUpdateSnapshots = false)
+@EnableSnapshotTests(forceUpdateSnapshots = false)
 public class SnapshotsTest {
 
     @Test
@@ -21,7 +22,7 @@ public class SnapshotsTest {
     @Test
     void testAsJsonStructureCompare(Snapshot snapshot) throws Exception {
         final Person myself = determinePerson();
-        final SnapshotResult snapshotResult = snapshot.assertThat(myself)
+        final SnapshotTestResult snapshotResult = snapshot.assertThat(myself)
                 .as(json)
                 .matchesSnapshotStructure();
         assertThat(snapshotResult.status()).isEqualTo(SnapshotStatus.ASSERTED);