+++ snapshot-tests-jaxb/src/test/java/de/skuzzle/test/snapshots/SnapshotsTest.java
@@ -10,21 +10,22 @@ import javax.xml.bind.annotation.XmlRootElement;
 import org.junit.jupiter.api.Test;
 
 import de.skuzzle.test.snapshots.SnapshotDsl.Snapshot;
+import de.skuzzle.test.snapshots.SnapshotTestResult.SnapshotStatus;
 
-@SnapshotAssertions(forceUpdateSnapshots = false)
+@EnableSnapshotTests(forceUpdateSnapshots = false)
 public class SnapshotsTest {
 
     @Test
     void testAsXmlTextCompare(Snapshot snapshot) throws Exception {
         final Person myself = determinePerson();
-        final SnapshotResult snapshotResult = snapshot.assertThat(myself).as(xml).matchesSnapshotText();
+        final SnapshotTestResult snapshotResult = snapshot.assertThat(myself).as(xml).matchesSnapshotText();
         assertThat(snapshotResult.status()).isEqualTo(SnapshotStatus.ASSERTED);
     }
 
     @Test
     void testAsXmlStructureCompare(Snapshot snapshot) throws Exception {
         final Person myself = determinePerson();
-        final SnapshotResult snapshotResult = snapshot.assertThat(myself).as(xml).matchesSnapshotStructure();
+        final SnapshotTestResult snapshotResult = snapshot.assertThat(myself).as(xml).matchesSnapshotStructure();
         assertThat(snapshotResult.status()).isEqualTo(SnapshotStatus.ASSERTED);
     }