+++ readme/RELEASE_NOTES.md
@@ -1,8 +1,6 @@
 [![Maven Central](https://img.shields.io/static/v1?label=MavenCentral&message=${project.version}&color=blue)](https://search.maven.org/artifact/${project.groupId}/${project.artifactId}/${project.version}/jar) [![JavaDoc](https://img.shields.io/static/v1?label=JavaDoc&message=${project.version}&color=orange)](http://www.javadoc.io/doc/${project.groupId}/${project.artifactId}/${project.version})
 
-* Refactor to multi module project
-* [#6](https://github.com/skuzzle/snapshot-tests/issues/6) Allow to specify explicit snapshot name
-* [#7](https://github.com/skuzzle/snapshot-tests/issues/7) Write snapshot information to headers (**breaking**)
+* [#11](https://github.com/skuzzle/snapshot-tests/issues/11) Rename `@SnapshotAssertions` to `@EnableSnapshotTests` (**breaking**)
 
 Maven Central coordinates for this release: